backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langgraph
langchain-google-genai
langsmith
dotenv
pyjwt